LPS-130662 Modularize
