+++ modules/apps/object/object-api/bnd.bnd
@@ -2,8 +2,8 @@ Bundle-Name: Liferay Object API
 Bundle-SymbolicName: com.liferay.object.api
 Bundle-Version: 1.0.0
 Export-Package:\
+	com.liferay.object.deployer,\
 	com.liferay.object.exception,\
-	com.liferay.object.graphql,\
 	com.liferay.object.model,\
 	com.liferay.object.service,\
 	com.liferay.object.service.persistence
