+++ modules/apps/object/object-api/src/main/resources/com/liferay/object/deployer/packageinfo
