+++ modules/apps/object/object-rest-api/bnd.bnd
@@ -1,0 +1,4 @@
+Bundle-Name: Liferay Object REST API
+Bundle-SymbolicName: com.liferay.object.rest.api
+Bundle-Version: 1.0.0
+Export-Package: com.liferay.object.rest.graphql
