+++ modules/apps/object/object-rest-api/build.gradle
@@ -1,0 +1,5 @@
+dependencies {
+	compileOnly group: "com.liferay.portal", name: "com.liferay.portal.kernel", version: "default"
+	compileOnly project(":apps:object:object-api")
+	compileOnly project(":apps:portal-odata:portal-odata-api")
+}
