+++ modules/apps/object/object-rest-api/src/main/resources/com/liferay/object/rest/graphql/packageinfo
@@ -1,0 +1,1 @@
+version 1.0.0
