+++ modules/apps/object/object-rest-impl/bnd.bnd
@@ -1,0 +1,3 @@
+Bundle-Name: Liferay Object REST Implementation
+Bundle-SymbolicName: com.liferay.object.rest.impl
+Bundle-Version: 1.0.0
