+++ modules/apps/object/object-rest-impl/build.gradle
@@ -1,0 +1,8 @@
+dependencies {
+	compileOnly group: "com.liferay.portal", name: "com.liferay.portal.kernel", version: "default"
+	compileOnly group: "org.osgi", name: "org.osgi.service.component.annotations", version: "1.3.0"
+	compileOnly group: "org.osgi", name: "osgi.core", version: "6.0.0"
+	compileOnly project(":apps:object:object-api")
+	compileOnly project(":apps:object:object-rest-api")
+	compileOnly project(":apps:portal-odata:portal-odata-api")
+}
