+++ modules/apps/object/object-rest-impl/src/main/java/com/liferay/object/rest/internal/deployer/ObjectDefinitionDeployerImpl.java
@@ -1,0 +1,64 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.object.rest.internal.deployer;
+
+import com.liferay.object.deployer.ObjectDefinitionDeployer;
+import com.liferay.object.model.ObjectDefinition;
+import com.liferay.object.rest.graphql.ObjectDefinitionGraphQL;
+import com.liferay.object.rest.internal.graphql.ObjectDefinitionGraphQLImpl;
+import com.liferay.object.service.ObjectFieldLocalService;
+import com.liferay.portal.kernel.util.HashMapDictionaryBuilder;
+
+import java.util.Collections;
+import java.util.List;
+
+import org.osgi.framework.BundleContext;
+import org.osgi.framework.ServiceRegistration;
+import org.osgi.service.component.annotations.Activate;
+import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.Reference;
+
+/**
+ * @author Brian Wing Shun Chan
+ */
+@Component(immediate = true, service = ObjectDefinitionDeployer.class)
+public class ObjectDefinitionDeployerImpl implements ObjectDefinitionDeployer {
+
+	public List<ServiceRegistration<?>> deploy(
+		ObjectDefinition objectDefinition) {
+
+		return Collections.singletonList(
+			_bundleContext.registerService(
+				ObjectDefinitionGraphQL.class,
+				new ObjectDefinitionGraphQLImpl(
+					objectDefinition,
+					_objectFieldLocalService.getObjectFields(
+						objectDefinition.getObjectDefinitionId())),
+				HashMapDictionaryBuilder.<String, Object>put(
+					"db.table.name", objectDefinition.getDBTableName()
+				).build()));
+	}
+
+	@Activate
+	protected void activate(BundleContext bundleContext) {
+		_bundleContext = bundleContext;
+	}
+
+	private BundleContext _bundleContext;
+
+	@Reference
+	private ObjectFieldLocalService _objectFieldLocalService;
+
+}
