+++ modules/apps/object/object-web/src/main/java/com/liferay/object/web/internal/application/list/ObjectDefinitionPanelApp.java
@@ -1,0 +1,69 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.object.web.internal.application.list;
+
+import com.liferay.application.list.BasePanelApp;
+import com.liferay.object.model.ObjectDefinition;
+import com.liferay.petra.string.StringPool;
+import com.liferay.portal.kernel.exception.PortalException;
+
+import java.util.Locale;
+
+import javax.portlet.PortletURL;
+
+import javax.servlet.http.HttpServletRequest;
+
+/**
+ * @author Marco Leo
+ * @author Brian Wing Shun Chan
+ */
+public class ObjectDefinitionPanelApp extends BasePanelApp {
+
+	public ObjectDefinitionPanelApp(ObjectDefinition objectDefinition) {
+		_objectDefinition = objectDefinition;
+	}
+
+	@Override
+	public String getKey() {
+		return super.getKey() + StringPool.POUND +
+			_objectDefinition.getObjectDefinitionId();
+	}
+
+	@Override
+	public String getLabel(Locale locale) {
+		return _objectDefinition.getName();
+	}
+
+	@Override
+	public String getPortletId() {
+		return _objectDefinition.getPortletId();
+	}
+
+	@Override
+	public PortletURL getPortletURL(HttpServletRequest httpServletRequest)
+		throws PortalException {
+
+		PortletURL portletURL = super.getPortletURL(httpServletRequest);
+
+		portletURL.setParameter(
+			"objectDefinitionId",
+			String.valueOf(_objectDefinition.getObjectDefinitionId()));
+
+		return portletURL;
+	}
+
+	private final ObjectDefinition _objectDefinition;
+
+}
