+++ modules/apps/object/object-web/src/main/java/com/liferay/object/web/internal/deployer/ObjectDefinitionDeployerImpl.java
@@ -1,0 +1,72 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.object.web.internal.deployer;
+
+import com.liferay.application.list.PanelApp;
+import com.liferay.application.list.constants.PanelCategoryKeys;
+import com.liferay.object.deployer.ObjectDefinitionDeployer;
+import com.liferay.object.model.ObjectDefinition;
+import com.liferay.object.web.internal.application.list.ObjectDefinitionPanelApp;
+import com.liferay.object.web.internal.portlet.ObjectDefinitionPortlet;
+import com.liferay.portal.kernel.util.HashMapDictionaryBuilder;
+
+import java.util.Arrays;
+import java.util.List;
+
+import javax.portlet.Portlet;
+
+import org.osgi.framework.BundleContext;
+import org.osgi.framework.ServiceRegistration;
+import org.osgi.service.component.annotations.Activate;
+import org.osgi.service.component.annotations.Component;
+
+/**
+ * @author Brian Wing Shun Chan
+ */
+@Component(immediate = true, service = ObjectDefinitionDeployer.class)
+public class ObjectDefinitionDeployerImpl implements ObjectDefinitionDeployer {
+
+	public List<ServiceRegistration<?>> deploy(
+		ObjectDefinition objectDefinition) {
+
+		return Arrays.asList(
+			_bundleContext.registerService(
+				PanelApp.class, new ObjectDefinitionPanelApp(objectDefinition),
+				HashMapDictionaryBuilder.<String, Object>put(
+					"panel.app.order:Integer", "300"
+				).put(
+					"panel.category.key", PanelCategoryKeys.CONTROL_PANEL_USERS
+				).build()),
+			_bundleContext.registerService(
+				Portlet.class, new ObjectDefinitionPortlet(),
+				HashMapDictionaryBuilder.<String, Object>put(
+					"com.liferay.portlet.display-category", "category.hidden"
+				).put(
+					"javax.portlet.display-name", objectDefinition.getName()
+				).put(
+					"javax.portlet.name", objectDefinition.getPortletId()
+				).put(
+					"javax.portlet.init-param.view-template", "/view.jsp"
+				).build()));
+	}
+
+	@Activate
+	protected void activate(BundleContext bundleContext) {
+		_bundleContext = bundleContext;
+	}
+
+	private BundleContext _bundleContext;
+
+}
