+++ modules/apps/object/object-web/src/main/resources/META-INF/resources/view.jsp
